+++ lib/hermes/tracers.rb
@@ -1,0 +1,5 @@
+module Hermes
+  module Tracers
+    autoload :DataDog, "hermes/tracers/data_dog"
+  end
+end